Move the render function from the SQL CLI to the python SDK so users only need to install the python SDK to use the render function.
