+++ python-sdk/src/astro/render/workflow_directory_parser.py
@@ -10,11 +10,51 @@ import yaml
 from airflow.models import BaseOperator
 from yaml.scanner import ScannerError
 
+from astro.constants import GENERATED_WORKFLOW_INCLUDE_DIRECTORY
+from astro.custom_backend.serializer import deserialize
+from astro.render.jinja import find_template_variables
 from astro.sql import LoadFileOperator
 from astro.sql.operators.transform import TransformOperator
-from sql_cli.constants import GENERATED_WORKFLOW_INCLUDE_DIRECTORY
-from sql_cli.operators.load_file import get_load_file_instance
-from sql_cli.utils.jinja import find_template_variables
+
+
+def build_load_file_operator_content(
+    operator_content: dict[str, Any],
+    operator_file_name: str,
+) -> dict[str, Any]:
+    """
+    Deserialize the File and Table classes.
+    :param operator_content: The dictionary of the operator content to be deserialized.
+    :param operator_file_name: The file name of the operator.
+    :returns: the operator content deserialized.
+    """
+    input_file_content = operator_content.pop("input_file")
+    input_file_content["class"] = "File"
+    input_file_content.setdefault("conn_id", None)
+    input_file_content.setdefault("is_dataframe", False)
+    input_file_content.setdefault("normalize_config", None)
+
+    operator_content["input_file"] = deserialize(input_file_content)
+
+    output_table_content = operator_content.pop("output_table")
+    output_table_content["class"] = "Table"
+    output_table_content.setdefault("metadata", {})
+    output_table_content.setdefault("name", operator_file_name)
+    output_table_content.setdefault("temp", False)
+
+    operator_content["output_table"] = deserialize(output_table_content)
+
+    return operator_content
+
+
+def get_load_file_instance(yaml_content: dict[str, Any], operator_file_name: str) -> LoadFileOperator:
+    """
+    Instantiate the LoadFileOperator.
+    :param yaml_content: The content of the load_file yaml.
+    :param operator_file_name: The file name of the operator.
+    :returns: an instance of the LoadFileOperator
+    """
+    operator_content: dict[str, Any] = build_load_file_operator_content(yaml_content, operator_file_name)
+    return LoadFileOperator(**operator_content)
 
 
 class WorkflowFile:
@@ -194,6 +234,7 @@ class SqlFile(WorkflowFile):
             "schema": self.metadata.get("schema"),
             "python_callable": lambda: (str(self.path), None),
             "sql": self.content,
+            "task_id": "transform_" + self.path.stem,
         }
         if airflow.__version__.startswith("2.2."):
             kwargs["op_args"] = []
