+++ python-sdk/tests/sql/operators/render_files/a.sql
@@ -1,0 +1,4 @@
+---
+conn_id: my_test_sqlite
+---
+select 1
