+++ python-sdk/tests/sql/operators/render_files/b.sql
@@ -1,0 +1,1 @@
+select * from {{ a }}
