+++ python-sdk/tests/sql/operators/render_files_with_load/a.sql
@@ -1,0 +1,4 @@
+---
+conn_id: my_test_sqlite
+---
+select * FROM {{ingest}}
