+++ python-sdk/tests/sql/operators/render_files_with_load/c.sql
@@ -1,0 +1,3 @@
+select * from {{ a }}
+union
+select * from {{ b }}
