+++ sql-cli/pyproject.toml
@@ -40,8 +40,7 @@ black = { version = "*", allow-prereleases = true }
 python-frontmatter = "^1.0.0"
 PyYAML = ">=5.4.1"
 python-dotenv = "^0.21.0"
-# Set the minimum version to v1.2.2 as the load_file operator relies on PR #1259 which was added in that release.
-astro-sdk-python = {extras = ["all"], version = "^1.2.2"}
+astro-sdk-python = {extras = ["all"], version = "^1.3.0.dev1"}
 apache-airflow = ">2.1"
 
 [tool.poetry.scripts]
@@ -55,6 +54,7 @@ pytest = "^7.1.3"
 pytest-cov = "^4.0.0"
 mypy = "^0.982"
 types-pyyaml = ">5.4.1"
+astro-sdk-python = {extras = ["all"], path="../python-sdk", develop=true}
 
 [tool.poetry.group.type_check]
 optional = true
@@ -62,6 +62,7 @@ optional = true
 [tool.poetry.group.type_check.dependencies]
 mypy = "^0.982"
 types-pyyaml = ">5.4.1"
+astro-sdk-python = {extras = ["all"], path="../python-sdk", develop=true}
 
 [build-system]
 requires = ["poetry-core"]
